feat(#685): moved useJwt hook to @hypha-platform/core
