+++ apps/web/src/app/[lang]/@aside/(.)profile/edit/page.tsx
@@ -2,7 +2,7 @@
 
 import { EditPersonSection } from '@hypha-platform/epics';
 import { useMe } from '@web/hooks/use-me';
-import { useJwt } from '@web/hooks/use-jwt';
+import { useJwt } from '@hypha-platform/core/client';
 import { useEditPersonOrchestrator } from '@core/people/client/hooks';
 import { SidePanel } from '../../_components/side-panel';
 import { LoadingBackdrop } from '@hypha-platform/ui/server';
