+++ apps/web/src/hooks/use-spaces-by-member-slug.ts
@@ -2,9 +2,8 @@
 
 import React from 'react';
 import useSWR from 'swr';
-import { useJwt } from './use-jwt';
 // TODO: #594 declare UI interface separately
-import { Space } from '@hypha-platform/core/client';
+import { Space, useJwt } from '@hypha-platform/core/client';
 
 type UseSpacesByMemberSlugReturn = {
   spaces: Space[];
