+++ packages/core/src/people/client/hooks/index.ts
@@ -1,1 +1,2 @@
 export { useEditPersonOrchestrator } from './useEditPersonOrchestrator';
+export { useJwt } from './useJwt';
