+++ packages/core/src/people/index.ts
@@ -1,2 +1,3 @@
 export * from './types';
 export * from './validation';
+export * from './client/hooks';
